chore(insights): restructure trends charts under insights/trends/
Reorganise the chart folders into the insights/trends/ structure with
per-component PascalCase directories:

  products/product_analytics/frontend/insights/trends/
    TrendsBarChart/
    TrendsLineChart/
    shared/

External importers (Trends.tsx, FunnelLineChart.tsx, funnelChartTransforms.ts)
updated to the new paths.

Generated-By: PostHog Code
Task-Id: 779fb86d-40c5-4855-bb6a-d2f16013824d
