+++ products/product_analytics/frontend/insights/trends/TrendsBarChart/TrendsBarChart.test.tsx
