+++ products/product_analytics/frontend/insights/trends/TrendsBarChart/handleTrendsBarAggregatedChartClick.ts
