+++ products/product_analytics/frontend/insights/trends/TrendsBarChart/trendsBarChartTransforms.test.ts
