+++ products/product_analytics/frontend/insights/trends/shared/AnnotationsLayer.tsx
