+++ products/product_analytics/frontend/insights/trends/shared/TrendsAlertOverlays.tsx
