+++ products/product_analytics/frontend/insights/trends/shared/anomalyPointsAdapter.test.ts
