+++ products/product_analytics/frontend/insights/trends/shared/goalLinesAdapter.ts
