+++ products/product_analytics/frontend/insights/trends/shared/handleTrendsChartClick.test.tsx
