+++ products/product_analytics/frontend/insights/trends/shared/trendsAxisFormat.ts
