+++ products/product_analytics/frontend/insights/trends/shared/trendsSeriesMeta.ts
